+++ ChangeRequestApproval/src/main/java/com/esewa/restchangerequestapproval/userdetails/service/UserService.java
@@ -1,8 +1,7 @@
 package com.esewa.restchangerequestapproval.userdetails.service;
 
-import com.esewa.restchangerequestapproval.security.user.User;
 import com.esewa.restchangerequestapproval.userdetails.userDto.UserResponseDto;
 
 public interface UserService {
-    UserResponseDto getUserById(long userId);
+    UserResponseDto getUserById(Integer userId);
 }
